refactor: extract property panel style actions
